% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{BIOSAMPLE_PACKAGES}
\alias{BIOSAMPLE_PACKAGES}
\title{An included table of BioSample Package information.}
\format{An object of class \code{data.frame} with 105 rows and 7 columns.}
\usage{
BIOSAMPLE_PACKAGES
}
\description{
See \code{\link{download_biosample_packages}} to get a fresh copy.
}
\keyword{datasets}
