% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{biosamples}
\alias{biosamples}
\title{NCBI BioSamples}
\description{
These functions help manage BioSamples spreadsheets and individual attribute
fields.
}
\details{
Notable functions:
\itemize{
\item \code{\link{build_biosamples_from_template}}: make a new spreadsheet from a
template name and existing metadata
\item \code{\link{write_biosamples}}: save a prepared biosamples spreadsheet to a
}
}
