% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{blank}
\alias{blank}
\title{Check for blank cells}
\usage{
blank(data)
}
\arguments{
\item{data}{vector of data to check.}
}
\value{
True if NA or empty, FALSE otherwise.
}
\description{
Check for blank cells
}
