% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{build_metadata}
\alias{build_metadata}
\title{Create blank metadata}
\usage{
build_metadata(sample_attrs, submission = NULL, col_pairs = NULL,
  constants = NULL)
}
\value{
data frame with SRA metadata columns
}
\description{
Create a new empty SRA metadata data frame with columns named and typed
accordingly.
}
