% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra_submission.R
\name{build_sra_submission}
\alias{build_sra_submission}
\title{Create BioSamples Attributes and SRA Metadata Tables}
\usage{
build_sra_submission(package_name, biosample_attrs,
  biosample_col_pairs = NULL, biosample_constants = NULL,
  metadata_attrs = NULL, metadata_col_pairs = NULL,
  metadata_constants = NULL, submission_accession = NULL)
}
\arguments{
\item{package_name}{name of BioSample Package for the template to use, as
either the display name (e.g., "MIGS: cultured bacteria/archaea,
human-associated; version 4.0") or the short name (e.g.,
"MIGS.ba.human-associated.4.0")}

\item{biosample_attrs}{data frame of existing sample metadata to draw from
for biological samples.  Any names given in the \code{biosample_col_pairs}
argument will be used to explicitly map column names from the existing data
frame to the new data frame.  Remaining columns with matching names will
also be used.  Other columns not in \code{biosample_col_pairs} or the
template's fields are ignored.}

\item{biosample_col_pairs}{named vector of column names in the existing data
frame with names set to column names in the new data frame.  Vector names
that don't match known column names signify custom columns to add.}

\item{biosample_constants}{vector of field names to match to constant values
for all samples.}

\item{metadata_attrs}{data frame of existing library metadata to draw from,
analagous to \code{biosample_attrs}.  If not specified,
\code{biosample_attrs} wil be used.}

\item{metadata_col_pairs}{named vector for metadata, as for
\code{biosample_col_pairs}.}

\item{metadata_constants}{named vector for metadata, as for
\code{biosample_constants}.}

\item{submission_accession}{the accession assigned by the submission portal
for the submission, like "SUB####".  Will be attached to the output data
frames as an attribute if given.}
}
\value{
list of data frames for BioSample Attributes and SRA Metadata.
}
\description{
Create all required SRA metadata in one go from the given data frames and
returns list of resulting data frames.  This encapsulates both
\code{\link{build_biosamples_from_template}} and
\code{\link{build_metadata}}.
}
\details{
If there is a one-to-one match between biological samples and sequencer
samples, \code{metadata_attrs} can be left out and \code{biosamples_attrs}
will be used for both spreadsheets.
}
