% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{check_uniqueness}
\alias{check_uniqueness}
\title{Check sample uniqueness from metadata}
\usage{
check_uniqueness(data)
}
\arguments{
\item{data}{data frame of BioSample attributes}
}
\value{
factor with levels corresponding to unique rows after excluding
  certain columns.
}
\description{
Per NCBI's rules: "You should have one BioSample for each specimen, and each
of your BioSamples must have differentiating information (excluding sample
name, title, bioproject accession and description). This check was
implemented to encourage submitters to include distinguishing information in
their samples. If the distinguishing information is in the sample name, title
or description, please recode it into an appropriate attribute, either one of
the predefined attributes or a custom attribute you define. If it is
necessary to represent true biological replicates as separate BioSamples, you
might add an 'aliquot' or 'replicate' attribute, e.g., 'replicate =
biological replicate 1', as appropriate. Note that multiple assay types,
e.g., RNA-seq and ChIP-seq data may reference the same BioSample if
appropriate."
}
