% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{chunk_submission}
\alias{chunk_submission}
\title{Chunk up submission data frames}
\usage{
chunk_submission(submission, rowsmax = 1000)
}
\description{
The SRA web interface will not allow a spreadsheet to be uploaded that
contains more than 1000 rows, for both BioSample Attributes and Metadata
(even though the documentation just says "1000 samples.")
https://www.ncbi.nlm.nih.gov/sra/docs/submitportal/
This will take a submission list pair of biosample/metadata data frames and
split each into its own list, keeping the number of rows below 1000 for all
output data frames.  The resulting spreadsheets will need to be uploaded in
separate submissions under the same BioProject.  The data frames here will be
organized so a given sample has all of its data in the same "chunk."
}
