% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{download_biosample_attributes}
\alias{download_biosample_attributes}
\title{Download BioSample Attributes XML}
\usage{
download_biosample_attributes(fp = NULL)
}
\arguments{
\item{fp}{file path to write to, or NULL to use a default within the package
directory.}
}
\value{
file path written to
}
\description{
Download an XML file of BioSample field attribute information and write to
disk.
}
