% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{download_biosample_packages}
\alias{download_biosample_packages}
\title{Download BioSample Package info XML}
\usage{
download_biosample_packages(fp = NULL)
}
\arguments{
\item{fp}{file path to write to, or NULL to use a default within the package
directory.}
}
\value{
file path written to
}
\description{
Download an XML file of BioSample package information and write to disk.
}
