% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{download_metadata}
\alias{download_metadata}
\title{Download Library Metadata Template}
\usage{
download_metadata(fp = NULL)
}
\arguments{
\item{fp}{file path to write to, or NULL to use a default within the package
directory.}
}
\value{
file path written to
}
\description{
Download the library metadata template as a TSV file.
}
