% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{dump_metadata}
\alias{dump_metadata}
\title{Download SRA BioSample Metadata}
\usage{
dump_metadata(dp = NULL)
}
\arguments{
\item{dp}{directory path to save files to.  If not specified, the extdata
subdirectory of the package location will be used.}
}
\value{
named vector of downloaded file paths.
}
\description{
Download all available BioSample package and template information and save to
a given directory.
}
