% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{field_descriptions}
\alias{field_descriptions}
\title{Get BioSample field descriptions}
\usage{
field_descriptions(fields)
}
\arguments{
\item{fields}{vector of field names, like \code{c("env_biome", "env_feature")}.}
}
\value{
named vector of descriptions, named by the fields given.
}
\description{
Given a vector of BioSample field (column) names, return SRA descriptions for
each.
}
