% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{fill_blanks}
\alias{fill_blanks}
\title{Fill blanks with NCBI-allowed keyword}
\usage{
fill_blanks(vec, data = NULL, value = "missing")
}
\arguments{
\item{vec}{either vector of column names (if data not NULL) or explicit data.}

\item{data}{data frame to use, if filling multiple columns.}

\item{value}{text to fill with; one of the allowed values in
\code{\link{BLANK_TYPES}}.}
}
\value{
data modified vector or data frame with blanks filled in.
}
\description{
Factors will be ignored since they have their own controlled vocabulary, and
biosample_accession and bioproject_accession columns will be ignored for data
frames as these are allowed to be left blank for all-in-one SRA submissions.
}
