% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fill_from_columns}
\alias{fill_from_columns}
\title{Fill in columns from reference}
\usage{
fill_from_columns(data_new, data_old, col_pairs = NULL)
}
\arguments{
\item{data_new}{data frame to modify.  All column names present here will be
checked for in \code{data_old}, enforcing data classes in \code{data_new}.}

\item{data_old}{data frame to pull from.  Column names here matching columns
in \code{data_new} will be used.}

\item{col_pairs}{character vector of column names in new data frame (vector
names) matched to column names in old data frame (vector values) to use.}
}
\value{
data frame of \code{data_new} values updated with matching columns
  from \code{data_old}.
}
\description{
Fill columns in one data frame from another, using data classes from the
destination data frame.  Optionally give a vector of mappings from one set of
column names to the other.
}
