% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FIXED_VOCABULARY}
\alias{FIXED_VOCABULARY}
\title{Metadata fields with restricted values}
\format{An object of class \code{list} of length 6.}
\usage{
FIXED_VOCABULARY
}
\description{
These are metadata field names that can only take on the values listed for
each name.
}
\keyword{datasets}
