% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdata.R
\name{has_zero_seqs}
\alias{has_zero_seqs}
\title{Report files containing no sequences}
\usage{
has_zero_seqs(filepaths)
}
\arguments{
\item{filepaths}{character vector of file paths to check}
}
\value{
logical vector, TRUE if file contains no sequences or FALSE otherwise
}
\description{
Given a character vector of file paths, produce a logical vector with TRUE
for files that contain no sequences and FALSE otherwise.  Supports FASTA and
FASTQ, as plain text, gzip-compressed, or bzip-compressed.
}
