% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{HTTP_SRV}
\alias{HTTP_SRV}
\title{NCBI server hostnames}
\format{An object of class \code{character} of length 2.}
\usage{
HTTP_SRV
}
\description{
NCBI server hostnames
}
\keyword{datasets}
