% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_template_types}
\alias{list_template_types}
\title{List template types}
\usage{
list_template_types()
}
\value{
vector of template type names.
}
\description{
List types of SRA templates recognized.
}
