% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_templates}
\alias{list_templates}
\title{List installed template files}
\usage{
list_templates(template_type = "biosample_attributes")
}
\arguments{
\item{template_type}{either "biosample_attributes" or "library_metadata".}
}
\value{
vector of template names, without file extension.
}
\description{
List names of templates installed in the package.  These are the spreadsheets
the SRA web interface gives you during a submission at the attributes and
metadata steps.
}
