% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{process_fixed_vocab}
\alias{process_fixed_vocab}
\title{Create factors for known fields with fixed vocabulary}
\usage{
process_fixed_vocab(data)
}
\arguments{
\item{data}{data frame to modify.}
}
\value{
data frame with recognized columns converted to factors.
}
\description{
For SRA fields in the given data frame known to have a strictly fixed
vocabulary (according to \code{FIXED_VOCABULARY}), convert each to a factor
with levels corresponding to the vocabulary.
}
