% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{read_biosample_attributes}
\alias{read_biosample_attributes}
\title{Read BioSample attributes list from file}
\usage{
read_biosample_attributes(fp = NULL)
}
\arguments{
\item{fp}{file path to read from, or NULL to use a default within the package
directory.}
}
\value{
nested list of BioSample field attributes.  Each named entry refers
  to a single field (column) used by one or more package templates.  Each
  entry contains "metadata" and "extra"; metadata stores basic information
  like a description of the field and what packages it's used in.  extra
  refers to additional per-metadata-item characteristics (currently only used
  for Packages).
}
\description{
Read in an XML file of BioSample field attributes and return as nested list
structure.  (This doesn't fit cleanly into a table like the packages
information does.)
}
