% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{read_biosample_packages}
\alias{read_biosample_packages}
\title{Read BioSample package information from file}
\usage{
read_biosample_packages(fp = NULL)
}
\arguments{
\item{fp}{file path to write to, or NULL to use a default within the package
directory.}
}
\value{
data frame of information for all available BioSample types.
}
\description{
Read in XML file of BioSample package information (including template
names) and return a data frame.
}
