% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqdata.R
\name{seqdata}
\alias{seqdata}
\title{Sequence Data Functions}
\description{
These functions help manage sequence data.  Currently there's just one public
function, to check for zero-sequence files.
}
\details{
Notable functions:
\itemize{
\item \code{\link{has_zero_seqs}}: check for empty fasta/fastq/fasta.gz/fastq.gz
}
}
