% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra_metadata.R
\name{sra_metadata}
\alias{sra_metadata}
\title{NCBI SRA Metadata}
\description{
These functions help manage SRA Metadata spreadsheets and individual
fields.
}
\details{
Notable functions:
\itemize{
\item \code{\link{build_metadata}}: make a new spreadsheet from existing
metadata and the standard template.
\item \code{\link{write_metadata}}: save a prepared metadata spreadsheet to a
file.
}
}
