% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra_submission.R
\name{sra_submission}
\alias{sra_submission}
\title{NCBI SRA Submissions}
\description{
These are high-level functions for handling SRA submission data (both
\link{biosamples} and \link{sra_metadata}).
}
\details{
Notable functions:
\itemize{
\item \code{\link{build_sra_submission}}: prepare both submission spreadsheets in
one step.
\item \code{\link{write_biosamples}}: save both prepared spreadsheets to files.
Also handles breaking up the spreadsheets if there would be too many rows
for the submission portal.
}
}
