% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables}
\alias{tables}
\title{NCBI Metadata Tables}
\description{
These functions handle NCBI metadata tables generally, and individual fields.
}
\details{
Notable functions:
\itemize{
\item \code{\link{read_table}}: read a TSV spreadsheet to a data frame
\item \code{\link{write_table}}: save a prepared spreadsheet to a file
\item \code{\link{validate_fields}}: check known NCBI metadata fields for
problems.
\item \code{\link{check_uniqueness}}: check for uniqueness of rows by
NCBI's rules
\item \code{\link{fill_blanks}}: fill in blank entries with one of the allowed
values
\item \code{\link{datemunge}}: Standardize date strings as YYY-MM-DD
}
}
