% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{tidy_optional_fields}
\alias{tidy_optional_fields}
\title{Remove blank optional fields}
\usage{
tidy_optional_fields(data)
}
\arguments{
\item{data}{data frame of metadata}
}
\value{
data frame with blank optional fields removed
}
\description{
Given a data frame (with optional_fields attribute), remove any columns that
are listed as optional and are entirely blank.  Call this after filling in
any optional fields that you do have information for to clear out the
remaining empty ones.
}
