% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Check SRA metadata for problems}
\usage{
validate_metadata(data, quiet = FALSE)
}
\arguments{
\item{data}{data frame of SRA fields}

\item{quiet}{skip calling \code{warning} for each problem?}
}
\value{
character vector of warnings
}
\description{
Check a data frame of SRA fieds (biosample attributes or library metadata)
for problems that may interfere with a submission.  Any problems found will
be raised as warnings (unless \code{quiet=TRUE}) and returned as a vector.
}
