% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{validate_submission}
\alias{validate_submission}
\title{Check both SRA spreadsheets for problems}
\usage{
validate_submission(submission, ...)
}
\arguments{
\item{submission}{list of data frames of SRA submission information}

\item{...}{any additional arguments to validate_fields}
}
\value{
list of character vector of warnings for each data frame
}
\description{
Check the biosample attributes or library metadata data frames for problems
using \code{\link{validate_fields}}.
}
