% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R, R/metadata.R
\name{write_biosamples}
\alias{write_biosamples}
\alias{write_metadata}
\title{Write SRA BioSample attributes or metadata to disk}
\usage{
write_biosamples(data, ...)

write_metadata(data)
}
\arguments{
\item{data}{data frame of SRA metadata}

\item{...}{additional arguments for \code{\link{write_sra_table}}}
}
\description{
These functions write SRA spreadsheets to the working directory by type
(BioSample attributes or library metadata) with the naming scheme
\code{<submission>/<submission>_<type>.tsv} if submission is defined or just
\code{<type>.tsv} otherwise.
}
\section{Functions}{
\itemize{
\item \code{write_biosamples}: Write SRA BioSample attributes to disk

\item \code{write_metadata}: Write SRA library metadata to disk
}}

