% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission.R
\name{write_submission}
\alias{write_submission}
\title{Write SRA submission spreadsheets to disk}
\usage{
write_submission(submission, ...)
}
\arguments{
\item{submission}{list of data frames of SRA submission information, such as
created by \code{\link{build_submission}}.}

\item{...}{additional arguments for \code{\link{write_sra_table}}}
}
\value{
list of vectors of file paths saved for each spreadsheet category
}
\description{
These functions write SRA spreadsheets to the working directory by type
(BioSample attributes and library metadata) with the naming scheme
\code{<submission>/<submission>_<type>_<part>.tsv}.  Each \code{part} splits
the submission spreadsheets so as to have fewer than 1000 rows in both, while
keeping entries for a given sample_name together between spreadsheet pairs.
\code{submission} will use the submission accession attribute on the data
frames, if present.  If either of these fields are not appliable they will be
excluded from the filename.  At its most basic this will simply save
\code{biosamples.tsv} and \code{metadata.tsv}.
}
