% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{build_biosamples_from_template}
\alias{build_biosamples_from_template}
\title{Create BioSamples Table}
\usage{
build_biosamples_from_template(template_name, sample_attrs,
  submission = NULL, col_pairs = NULL, constants = NULL)
}
\arguments{
\item{template_name}{name of submission type (e.g.,
"MIMS.me.human-associated.4.0")}

\item{sample_attrs}{data frame of existing sample metadata to draw from.  Any
names given in the \code{col_pairs} argument will be used to explicitly map
column names from the existing data frame to the new data frame.  Remaining
columns with matching names will also be used.  Other columns not in
\code{col_pairs} or the template's fields are ignored.}

\item{col_pairs}{named vector of column names in the existing data frame with
names set to column names in the new data frame.  Vector names that don't
match known column names signify custom columns to add.}

\item{constants}{vector of field names to match to constant values for all
samples.}
}
\value{
data frame with SRA BioSample attributes defined.
}
\description{
Create a new SRA BioSamples table using a named template and existing sample
attributes.
}
