% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{download_biosample_packages}
\alias{download_biosample_packages}
\title{Download BioSample Package Information}
\usage{
download_biosample_packages()
}
\value{
data frame of information for all available BioSample types.
}
\description{
Download an XML file of BioSample Package information (including template
names) and return a data frame.
}
