% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{download_template}
\alias{download_template}
\title{Download BioSample Template}
\usage{
download_template(package_name, fp = NULL)
}
\arguments{
\item{package_name}{name of BioSample Package for the template to download,
(e.g., "MIMS.me.human-associated.4.0")}

\item{fp}{file path to write to, or NULL to use a default within the package
directory.}
}
\value{
file path written to
}
\description{
Download a BioSample template as a TSV file.
}
