% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosamples.R
\name{download_template}
\alias{download_template}
\title{Download BioSample Template}
\usage{
download_template(package_name)
}
\value{
text of TSV template
}
\description{
Download a BioSample template as a TSV file and return the text.
}
