% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fill_blanks}
\alias{fill_blanks}
\title{Fill blanks with SRA keyword}
\usage{
fill_blanks(vec, data = NULL, value = "missing")
}
\arguments{
\item{vec}{either vector of column names (if data not NULL) or explicit data.}

\item{data}{data frame to use, if filling multiple columns.}

\item{value}{text to fill with; one of the allowed values in
\code{BLANK_TYPES}.}
}
\value{
data modified vector or data frame with blanks filled in.
}
\description{
Fill blanks with SRA keyword
}
