% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{read_template}
\alias{read_template}
\title{Read SRA table template}
\usage{
read_template(nm)
}
\arguments{
\item{template_name}{name of submission type (e.g.,
"MIMS.me.human-associated.4.0")}
}
\value{
data frame with SRA field names and types set by template.
}
\description{
load an SRA template by name (e.g., "MIMS.me.human-associated.4.0") and
return as a data frame.
}
