% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srahelper.R
\docType{package}
\name{srahelper-package}
\alias{srahelper}
\alias{srahelper-package}
\title{SRA Helper}
\description{
A jumble of helper functions to assist with preparing NCBI data submissions
and working with the associated files.
}
\details{
Some summaries by topic:
\itemize{
\item \link{biosamples}
}
}
\author{
\strong{Maintainer}: Jesse Connell \email{ancon@upenn.edu}

}
