% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{write_sra_table}
\alias{write_sra_table}
\alias{read_sra_table}
\title{Read and write SRA metadata to disk}
\usage{
write_sra_table(data, fp, ...)

read_sra_table(fp, ...)
}
\arguments{
\item{data}{data frame of SRA metadata}

\item{fp}{file path to save text}

\item{...}{additional arguments for \code{utils::write.table}}
}
\description{
These functions read and write to/from a data frame of SRA metadata (a
BioSamples spreadsheet, library prep metadata, or run info) to/from a file
in TSV format.
}
\section{Functions}{
\itemize{
\item \code{write_sra_table}: Write SRA metadata to disk

\item \code{read_sra_table}: Read SRA metadata from disk
}}

